+++ lib/cmdx/log_formatters/line.rb
@@ -5,7 +5,7 @@ module CMDx
     class Line
 
       def call(severity, time, progname, message)
-        time    = Utils::DatetimeFormatter.call(time.utc)
+        time    = Utils::LogTimestamp.call(time.utc)
         message = message.to_h.map { |k, v| "#{k}=#{v}" }.join(" ") if message.is_a?(Result)
 
         "#{severity[0]}, [#{time} ##{Process.pid}] #{severity} -- #{progname || 'CMDx'}: #{message}\n"